# src/docs/openapi/swagger.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "TRTL apps",
    "description" : "This is the API for apps to interact with the TRTL apps service.  You can find out more about TRTL apps at [http://trtlapps.io](http://trtlapps.io).",
    "version" : "1.0.0"
  },
  "servers" : [ {
    "url" : "https://trtlapps.io/api"
  } ],
...
  "paths" : {
    "/{appId}/accounts" : {
      "post" : {
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "app" ],
        "summary" : "Creates a new account",
        "description" : "Returns the newly created account object",
        "parameters" : [ {
          "name" : "appId",
          "in" : "path",
          "description" : "ID of the app creating the account",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Account"
                },
                "example" : {
                  "id" : "8RgwiWmgiYKQlUHWGaTW",
                  "appId" : "84U0IRP0rdC57AHfwRQc",
                  "balanceLocked" : 0,
                  "balanceUnlocked" : 0,
                  "createdAt" : 1568917025296,
                  "deleted" : false,
                  "paymentId" : "6a8db2c83a34f29275d3cfad7100944168c46fd5d43e074aa038e18a0410c7dd",
                  "depositAddress" : "TRTLuxsFnkbHRdfoKCFc8KJ6Rzvp1TdqLHdJctadbENjAZx5mhgXqQXA6kNfcuNnap94kdBUwnfvPHvh82YufKFmHbXLoKCzfHLXTXbsQxzqy5KHyHXTQZbtEHnAsorKPqkweHzDRcRyM28g6jJMQjBoocYqsjtR3G5n1ipuDpn6VbqZQWnQ1D22qvB",
                  "depositQrCode" : "https://chart.googleapis.com/chart?cht=qr&chs=256x256&chl=turtlecoin://TRTLuxsFnkbHRdfoKCFc8KJ6Rzvp1TdqLHdJctadbENjAZx5mhgXqQXA6kNfcuNnap94kdBUwnfvPHvh82YufKFmHbXLoKCzfHLXTXbsQxzqy5KHyHXTQZbtEHnAsorKPqkweHzDRcRyM28g6jJMQjBoocYqsjtR3G5n1ipuDpn6VbqZQWnQ1D22qvB"
                }
              }
            }
          },
          "401" : {
            "$ref" : "#/components/responses/UnauthorizedError"
          },
          "500" : {
            "description" : "Server error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ServiceError"
                }
              }
            }
          }
        }
      }
    },
    "/{appId}/accounts/{accountId}" : {
      "get" : {
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "app" ],
        "summary" : "Gets the Account with the given id",
        "description" : "Returns an Account object",
        "parameters" : [ {
          "name" : "appId",
          "in" : "path",
          "description" : "Your app ID.",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "accountId",
          "in" : "path",
          "description" : "The ID of the account to retrieve.",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Account"
                },
                "example" : {
                  "id" : "8RgwiWmgiYKQlUHWGaTW",
                  "appId" : "84U0IRP0rdC57AHfwRQc",
                  "balanceLocked" : 0,
                  "balanceUnlocked" : 0,
                  "createdAt" : 1568917025296,
                  "deleted" : false,
                  "paymentId" : "6a8db2c83a34f29275d3cfad7100944168c46fd5d43e074aa038e18a0410c7dd",
                  "depositAddress" : "TRTLuxsFnkbHRdfoKCFc8KJ6Rzvp1TdqLHdJctadbENjAZx5mhgXqQXA6kNfcuNnap94kdBUwnfvPHvh82YufKFmHbXLoKCzfHLXTXbsQxzqy5KHyHXTQZbtEHnAsorKPqkweHzDRcRyM28g6jJMQjBoocYqsjtR3G5n1ipuDpn6VbqZQWnQ1D22qvB",
                  "depositQrCode" : "https://chart.googleapis.com/chart?cht=qr&chs=256x256&chl=turtlecoin://TRTLuxsFnkbHRdfoKCFc8KJ6Rzvp1TdqLHdJctadbENjAZx5mhgXqQXA6kNfcuNnap94kdBUwnfvPHvh82YufKFmHbXLoKCzfHLXTXbsQxzqy5KHyHXTQZbtEHnAsorKPqkweHzDRcRyM28g6jJMQjBoocYqsjtR3G5n1ipuDpn6VbqZQWnQ1D22qvB"
                }
              }
            }
          },
          "400" : {
            "$ref" : "#/components/responses/BadRequest"
          },
          "401" : {
            "$ref" : "#/components/responses/UnauthorizedError"
          },
          "500" : {
            "description" : "Server error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ServiceError"
                }
              }
            }
          }
        }
      }
    },
    "/service/validateaddress" : {
      "get" : {
        "tags" : [ "service" ],
        "summary" : "Check address validity",
        "description" : "Checks if the supplied address is valid.",
        "parameters" : [ {
          "name" : "address",
          "in" : "query",
          "description" : "The address you want to validate.",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "allowIntegrated",
          "in" : "query",
          "description" : "Should integrated address be included as valid.",
          "required" : true,
          "schema" : {
            "type" : "boolean"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "isValid" : {
                      "type" : "boolean"
                    }
                  }
                },
                "example" : {
                  "isValid" : true
                }
              }
            }
          },
          "400" : {
            "$ref" : "#/components/responses/BadRequest"
          },
          "500" : {
            "description" : "Server error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ServiceError"
                }
              }
            }
          }
        }
      }
    },
    "/{appId}/accounts/{accountId}/withdrawaddress" : {
      "put" : {
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "app" ],
        "summary" : "Set an account's withdraw address",
        "description" : "Set the account withdraw address. Note that integrated addresses are not allowed. Returns the newly set address if the operation succeeded.",
        "parameters" : [ {
          "name" : "appId",
          "in" : "path",
          "description" : "Your app ID.",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "accountId",
          "in" : "path",
          "description" : "ID of the account.",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "required" : true,
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "object",
                "properties" : {
                  "address" : {
                    "type" : "string"
                  }
                }
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "Successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "address" : {
                      "type" : "string"
                    }
                  }
                },
                "example" : {
                  "address" : "TRTLv3wHDVyjcHXV2fHQ3cfSjjpCck9YjEEAw4jaAfn3Djrrz63bPCtWJNPW3CJhbg2BmnNidHMEo29geJc9QmXKNroz8NgLUc7"
                }
              }
            }
          },
          "400" : {
            "$ref" : "#/components/responses/BadRequest"
          },
          "500" : {
            "description" : "Server error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ServiceError"
                }
              }
            }
          }
        }
      }
    },
    "/{appId}/deposits" : {
      "get" : {
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "app" ],
        "summary" : "Gets the newest deposits of the supplied accountID",
        "description" : "Returns a list of Deposits",
        "parameters" : [ {
          "name" : "appId",
          "in" : "path",
          "description" : "Your app ID.",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "accountId",
          "in" : "query",
          "description" : "The account ID of the deposits owner.",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "limit",
          "in" : "query",
          "description" : "The amount of deposits to fetch.",
          "schema" : {
            "type" : "integer",
            "minimum" : 1,
            "maximum" : 100
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ArrayOfDeposits"
                }
              }
            }
          },
          "400" : {
            "$ref" : "#/components/responses/BadRequest"
          },
          "401" : {
            "$ref" : "#/components/responses/UnauthorizedError"
          },
          "500" : {
            "description" : "Server error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ServiceError"
                }
              }
            }
          }
        }
      }
    },
    "/{appId}/deposits/{depositId}" : {
      "get" : {
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "app" ],
        "summary" : "Gets the deposit with the given ID",
        "description" : "Returns a Deposit object",
        "parameters" : [ {
          "name" : "appId",
          "in" : "path",
          "description" : "Your app ID.",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "depositId",
          "in" : "path",
          "description" : "The ID of the deposit to retrieve.",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Deposit"
                },
                "example" : {
                  "id" : "Ehr3Zzfkvc01Ir3n0nMN",
                  "appId" : "J3irXkH26gE4pZHYyWWW",
                  "accountId" : "8RgwiWmgiYKQlUHWGaTW",
                  "paymentId" : "fa8ecc78c8d3a56895a102d30b0dea629bb75075b57869fb0395b56f48a2ae21",
                  "depositAddress" : "TRTLv3wHDVyjcHXV2fHQ3cfSjjpCck9YjEEAw4jaAfn3Djrrz63bPCtWJNPW3CJhbg2BmnNidHMEo29geJc9QmXKNroz8NgLUc7",
                  "integratedAddress" : "TRTLuz6RCHAHdBfSy6kBC6HG3CpR9PpZe94a3bGdZuzBHxWc1GGniGv9tzdJ7JpvV5A53V8GoBCTvHRjsvQqwgRwHHrfsxfrDoVjcHXV2fHQ3cfSjjpCck9YjEEAw4jaAfn3Djrrz63bPCtWJNPW3CJhbg2BmnNidHMEo29geJc9QmXKNroz8JtDGhG",
                  "amount" : 4200,
                  "txHash" : "6450ebe98e9387eab370d8394ab31b9a30822bd43c97737225ff8e0193abb6c8",
                  "createdDate" : 1573251262062,
                  "blockHeight" : 2086619,
                  "status" : "completed",
                  "accountCredited" : true,
                  "cancelled" : false,
                  "lastUpdate" : 1573251954262
                }
              }
            }
          },
          "400" : {
            "$ref" : "#/components/responses/BadRequest"
          },
          "401" : {
            "$ref" : "#/components/responses/UnauthorizedError"
          },
          "500" : {
            "description" : "Server error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ServiceError"
                }
              }
            }
          }
        }
      }
    },
    "/{appId}/transfers" : {
      "post" : {
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "app" ],
        "summary" : "Transfer funds from the sender to 1 or more receivers",
        "description" : "Transfers funds from the sender to the array of recipients. Returns a Transfer object if the operation succeeded.",
        "parameters" : [ {
          "name" : "appId",
          "in" : "path",
          "description" : "Your app ID.",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "required" : true,
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "object",
                "properties" : {
                  "senderId" : {
                    "type" : "string",
                    "example" : "hUMUXEFN9amjQjXN9Km7",
                    "description" : "ID of the account sending the funds"
                  },
                  "recipients" : {
                    "type" : "array",
                    "items" : {
                      "$ref" : "#/components/schemas/Recipient"
                    },
                    "example" : [ {
                      "accountId" : "DawR7cEvQjEWMBVmMkkn",
                      "amount" : 300
                    }, {
                      "accountId" : "rwszORa1qaSXK0RbZ7F5",
                      "amount" : 200
                    } ]
                  }
                }
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Transfer"
                },
                "example" : {
                  "transferId" : "apHo3ySWJAZy6KwH704C",
                  "appId" : "tX0vd7xGAE9aUrKF6oel",
                  "senderId" : "hUMUXEFN9amjQjXN9Km7",
                  "recipients" : [ {
                    "accountId" : "DawR7cEvQjEWMBVmMkkn",
                    "amount" : 300
                  }, {
                    "accountId" : "rwszORa1qaSXK0RbZ7F5",
                    "amount" : 200
                  } ],
                  "timestamp" : 1575137353275
                }
              }
            }
          },
          "400" : {
            "$ref" : "#/components/responses/BadRequest"
          },
          "401" : {
            "$ref" : "#/components/responses/UnauthorizedError"
          },
          "500" : {
            "description" : "Server error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ServiceError"
                }
              }
            }
          }
        }
      }
    },
    "/{appId}/transfers/{transferId}" : {
      "get" : {
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "app" ],
        "summary" : "Gets the transfer with the given transfer ID",
        "description" : "Returns a Transfer object",
        "parameters" : [ {
          "name" : "appId",
          "in" : "path",
          "description" : "Your app ID.",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "transferId",
          "in" : "path",
          "description" : "The ID of the transfer to retrieve.",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Transfer"
                },
                "example" : {
                  "id" : "apHo3ySWJAZy6KwH704C",
                  "appId" : "tX0vd7xGAE9aUrKF6oel",
                  "senderId" : "hUMUXEFN9amjQjXN9Km7",
                  "recipients" : [ {
                    "accountId" : "DawR7cEvQjEWMBVmMkkn",
                    "amount" : 300
                  }, {
                    "accountId" : "rwszORa1qaSXK0RbZ7F5",
                    "amount" : 200
                  } ],
                  "timestamp" : 1575137353275
                }
              }
            }
          },
          "400" : {
            "$ref" : "#/components/responses/BadRequest"
          },
          "401" : {
            "$ref" : "#/components/responses/UnauthorizedError"
          },
          "500" : {
            "description" : "Server error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ServiceError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "securitySchemes" : {
      "bearerAuth" : {
        "type" : "http",
        "scheme" : "bearer"
      }
    },
    "responses" : {
      "UnauthorizedError" : {
        "description" : "Access token is missing or invalid",
        "content" : {
          "application/json" : {
            "schema" : {
              "$ref" : "#/components/schemas/ServiceError"
            }
          }
        }
      },
      "BadRequest" : {
        "description" : "Bad request",
        "content" : {
          "application/json" : {
            "schema" : {
              "$ref" : "#/components/schemas/ServiceError"
            }
          }
        }
      }
    },
    "schemas" : {
      "Account" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string"
          },
          "appId" : {
            "type" : "string"
          },
          "paymentId" : {
            "type" : "string"
          },
          "depositAddress" : {
            "type" : "string"
          },
          "depositQrCode" : {
            "type" : "string"
          },
          "withdrawAddress" : {
            "type" : "string"
          },
          "balanceUnlocked" : {
            "type" : "integer"
          },
          "balanceLocked" : {
            "type" : "integer"
          },
          "createdAt" : {
            "type" : "integer"
          },
          "deleted" : {
            "type" : "boolean"
          }
        }
      },
      "ArrayOfDeposits" : {
        "type" : "array",
        "items" : {
          "$ref" : "#/components/schemas/Deposit"
        }
      },
...
      "Recipient" : {
        "type" : "object",
        "properties" : {
          "accountId" : {
            "type" : "string"
          },
          "amount" : {
            "type" : "integer"
          }
        }
      },
...
      "ServiceError" : {
        "type" : "object",
        "properties" : {
          "errorCode" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          }
        }
      }
    }
  }
}